Modify the request to handle errors gracefully
